agent/pyproject.toml
[project]
name = "truliv-agent"
version = "0.1.0"
description = "Truliv Voice AI Agent powered by LiveKit"
requires-python = ">=3.10,<3.14"
dependencies = [
    # LiveKit core
    "livekit-agents[silero,turn-detector,google,deepgram,cartesia,sarvam]~=1.3",
    # Config
    "python-dotenv",
    # HTTP clients
    "httpx",
    "aiohttp",
    "requests",
    # MongoDB
    "motor",
    "tenacity",
    # Google Sheets
    "gspread",
    "pandas",
    "google-auth",
    # LeadSquared / LLM for zero-deposit tool
    "langchain-core",
    "langchain-google-genai",
    # Logging
    "loguru",
]
